[CBO-67] Added ruff and tox support to the project and fixed linting errors.
