+++ backend/app/db/database.py
@@ -23,27 +23,24 @@ SOFTWARE AND ACCOMPANYING DOCUMENTATION, IF ANY, PROVIDED HEREUNDER IS PROVIDED
 ENHANCEMENTS, OR MODIFICATIONS.
 """
 
-from app.core.config_manager import settings
+from backend.app.core.config_manager import settings
+from sqlalchemy import create_engine
+from sqlalchemy.orm import Session, declarative_base, sessionmaker
+
+# Create SQLAlchemy engine
+engine = create_engine(settings.SQLALCHEMY_DATABASE_URI)
+
+# Create session factory
+SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
+
+# Create declarative base
+Base = declarative_base()
 
 
-def main() -> None:
-    """Print out key settings to verify configuration manager is working correctly."""
-    
-    print('Verifying configuration settings...')
-    print(f'PROJECT_NAME: {settings.PROJECT_NAME}')
-    print(f'VERSION: {settings.VERSION}')
-    print(f'ENVIRONMENT: {settings.ENVIRONMENT}')
-    print(f'APP_ENV: {settings.APP_ENV}')
-    print(f'DATABASE_URL: {settings.DATABASE_URL}')
-    print(f'AWS_REGION: {settings.AWS_REGION}')
-    print(f'BEDROCK_MODEL_ID: {settings.BEDROCK_MODEL_ID}')
-    print(f'BEDROCK_KNOWLEDGE_BASE_ID: {settings.BEDROCK_KNOWLEDGE_BASE_ID}')
-    print(f'LANGCHAIN_PROJECT: {settings.LANGCHAIN_PROJECT}')
-    print(f'LOGGING_LEVEL: {settings.LOGGING_LEVEL}')
-    print(f'LOG_TO_FILE: {settings.LOG_TO_FILE}')
-    print(f'LOGGING_LOCATION: {settings.LOGGING_LOCATION}')
-    print('Configuration verification complete.')
-
-
-if __name__ == '__main__':
-    main()
+# Dependency to get DB session
+def get_db() -> Session:
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
