+++ requirements.txt
@@ -9,3 +9,7 @@ psycopg2-binary==2.9.9
 # Pydantic
 pydantic>=2.7.4,<3.0.0
 pydantic-settings>=2.4.0,<3.0.0
+
+# Linting and Formatting
+ruff==0.5.0  # Replaces flake8
+pytest-ruff==0.2.1 # Replaces pytest-flake8
