+++ tox.ini
@@ -1,0 +1,22 @@
+[tox]
+minversion = 4.11.4
+envlist = lint
+isolated_build = True
+skipsdist = True
+
+[testenv]
+passenv = *
+deps =
+    -r{toxinidir}/requirements.txt
+commands =
+    # Run setup first
+    python -c "print('Running in environment: {envname}')"
+    # We'll use specialized environments for tests
+
+[testenv:lint]
+description = Run linting checks with ruff
+deps =
+    ruff==0.5.0
+commands =
+    ruff check {toxinidir}/backend {toxinidir}/frontend
+    ruff format --check {toxinidir}/backend {toxinidir}/frontend
